requirements.txt
streamlit
pymupdf
langchain
langchain-community
langchain-openai
openai
python-dotenv
faiss-cpu
tiktoken
